[CSL-1994] Move binaryTests to CborSpec
