+++ lib/cardano-sl.cabal
@@ -260,11 +260,7 @@ test-suite cardano-test
                        -- Standard module with some magic
                        Spec
 
-                       -- Something
-                       Test.Pos.BinarySpec
-
                        -- Pos.Block testing
-                       Test.Pos.Block.Identity.BinarySpec
                        Test.Pos.Block.Identity.SafeCopySpec
 
                        -- Core
@@ -278,7 +274,6 @@ test-suite cardano-test
 
                        -- Ssc
                        Test.Pos.Ssc.ComputeSharesSpec
-                       Test.Pos.Ssc.Identity.BinarySpec
                        Test.Pos.Ssc.Identity.SafeCopySpec
                        Test.Pos.Ssc.SeedSpec
                        Test.Pos.Ssc.Toss.BaseSpec
@@ -287,23 +282,16 @@ test-suite cardano-test
 
                        -- Everything else
                        Test.Pos.Cbor.CborSpec
-                       Test.Pos.Communication.Identity.BinarySpec
                        Test.Pos.ConstantsSpec
                        Test.Pos.CryptoSpec
-                       Test.Pos.Delegation.Identity.BinarySpec
-                       Test.Pos.DHT.Identity.BinarySpec
                        Test.Pos.Genesis.CanonicalSpec
                        Test.Pos.MerkleSpec
                        Test.Pos.Slotting.TypesSpec
-                       Test.Pos.Slotting.BinarySpec
                        Test.Pos.Txp.CoreSpec
-                       Test.Pos.Txp.Identity.BinarySpec
                        Test.Pos.Txp.Toil.UtxoSpec
                        Test.Pos.Types.BlockSpec
-                       Test.Pos.Types.Identity.BinarySpec
                        Test.Pos.Types.Identity.SafeCopySpec
                        Test.Pos.Types.Identity.ShowReadSpec
-                       Test.Pos.Update.Identity.BinarySpec
                        Test.Pos.Update.Identity.SafeCopySpec
                        Test.Pos.Update.MemStateSpec
                        Test.Pos.Update.PollSpec
