+++ lib/test/Test/Pos/CryptoSpec.hs
@@ -24,10 +24,9 @@ import           Pos.SafeCopy ()
 import           Pos.Ssc ()
 
 import           Pos.Util.QuickCheck.Property (qcIsLeft, (.=.))
-import           Test.Pos.Cbor.CborSpec (U)
 import           Test.Pos.Configuration (withDefConfiguration, withDefInfraConfiguration)
-import           Test.Pos.Helpers (binaryEncodeDecode, binaryTest, msgLenLimitedTest,
-                                   safeCopyEncodeDecode, safeCopyTest, serDeserId)
+import           Test.Pos.Helpers (binaryEncodeDecode, msgLenLimitedTest, safeCopyEncodeDecode,
+                                   safeCopyTest, serDeserId)
 
 {-# ANN module ("HLint: ignore Reduce duplication" :: Text) #-}
 
@@ -80,37 +79,6 @@ spec = withDefInfraConfiguration $ withDefConfiguration $ describe "Crypto" $ do
                  passphraseChangeLeavesAddressUnmodified
 
         describe "Identity testing" $ do
-            describe "Bi instances" $ do
-                binaryTest @Crypto.SecretKey
-                binaryTest @Crypto.PublicKey
-                binaryTest @(Crypto.Signature ())
-                binaryTest @(Crypto.Signature U)
-                binaryTest @(Crypto.ProxyCert Int32)
-                binaryTest @(Crypto.ProxySecretKey Int32)
-                binaryTest @(Crypto.ProxySecretKey U)
-                binaryTest @(Crypto.ProxySignature Int32 Int32)
-                binaryTest @(Crypto.ProxySignature U U)
-                binaryTest @(Crypto.Signed Bool)
-                binaryTest @(Crypto.Signed U)
-                binaryTest @Crypto.RedeemSecretKey
-                binaryTest @Crypto.RedeemPublicKey
-                binaryTest @(Crypto.RedeemSignature Bool)
-                binaryTest @(Crypto.RedeemSignature U)
-                binaryTest @Crypto.Threshold
-                binaryTest @Crypto.VssPublicKey
-                binaryTest @Crypto.PassPhrase
-                binaryTest @Crypto.VssKeyPair
-                binaryTest @Crypto.Secret
-                binaryTest @Crypto.DecShare
-                binaryTest @Crypto.EncShare
-                binaryTest @Crypto.SecretProof
-                binaryTest @Crypto.HDAddressPayload
-                binaryTest @(Crypto.AbstractHash Blake2b_224 U)
-                binaryTest @(Crypto.AbstractHash Blake2b_256 U)
-                binaryTest @(AsBinary Crypto.VssPublicKey)
-                binaryTest @(AsBinary Crypto.Secret)
-                binaryTest @(AsBinary Crypto.DecShare)
-                binaryTest @(AsBinary Crypto.EncShare)
             describe "SafeCopy instances" $ do
                 safeCopyTest @Crypto.SecretKey
                 safeCopyTest @Crypto.PublicKey
